refactor: Delete commented commented
I would like to improve upon the computer logic - to make it actively
try to win, instead of randomly choosing tiles for a move.

This can be done by writing an algorithm to determine if there is a winning move,
and for the computer to make that move if it can, else it will block winning moves.